Move definition of numeric_types into pyomo.common
