+++ pyomo/common/numeric_types.py
